added exercicio 06 and fixed exercicio 05

+++ Python/exercicio-05/solucao.py
@@ -1,6 +1,6 @@
 def calcular_media():
-    prova1 = int(input('Qual foi a nota da prova 1 ?'))
-    prova2 = int(input('Qual foi a nota da prova 2 ?'))
+    prova1 = float(input('Qual foi a nota da prova 1 ?'))
+    prova2 = float(input('Qual foi a nota da prova 2 ?'))
     media = (prova1 + prova2) / 2
     print(f'A média entre {prova1} e {prova2} é {media}')
     input('Pressione Enter Para Sair...')
